v3.4.11 - fixed sub redefined warning
The initial version of this package was developed using an old version
of Path::Class before it exported the 'dir' subroutine automatically, so
on newer versions, a sub redefined error would be thrown. To solve this,
the use was simply changed to a require.

Note that version '3.4.11' is actually the correct sub-version of ExtJS,
making this doable without breaking the construct of keeping the CPAN
version tied to the ExtJS version... If however another change were to
be needed, this would not be so (fingers crossed nothing else will be
needed!)

Also added a proper Changes file